+++ lib/interpreters/split_clustering.rb
@@ -28,6 +28,23 @@ require './lib/output_helper.rb'
 
 module Interpreters
   class SplitClustering
+    # Relations that link multiple services to be on the same route
+    LINKING_RELATIONS = %i[
+      order
+      same_route
+      sequence
+      shipment
+    ].freeze
+    # Relations that force multiple services/vehicles to stay in the same VRP
+    FORCING_RELATIONS = %i[
+      maximum_day_lapse
+      maximum_duration_lapse
+      meetup
+      minimum_day_lapse
+      minimum_duration_lapse
+      vehicle_trips
+    ].freeze
+
     # TODO: private method
     def self.split_clusters(service_vrp, job = nil, &block)
       vrp = service_vrp[:vrp]
@@ -124,27 +141,33 @@ module Interpreters
       vrp = service_vrp[:vrp]
       if service_vrp[:split_level].nil?
         empties_or_fills = vrp.services.select{ |s| s.quantities.any?(&:fill) || s.quantities.any?(&:empty) }
-        depot_ids = vrp.vehicles.flat_map{ |vehicle| [vehicle.start_point_id, vehicle.end_point_id].compact }.uniq
 
         !vrp.scheduling? &&
           vrp.preprocessing_max_split_size &&
           vrp.vehicles.size > 1 &&
           (vrp.resolution_vehicle_limit.nil? || vrp.resolution_vehicle_limit > 1) &&
-          (vrp.shipments.size + vrp.services.size - empties_or_fills.size) > vrp.preprocessing_max_split_size &&
-          vrp.shipments.all?{ |s| depot_ids.include?(s.pickup.point_id) || depot_ids.include?(s.delivery.point_id) }
+          (vrp.services.size - empties_or_fills.size) > vrp.preprocessing_max_split_size &&
+          vrp.shipments.empty? # Clustering supports Shipment only as Relation  TODO: delete this check when Model::Shipment is removed
       else
         ss_data = service_vrp[:split_solve_data]
-        current_vehicles = ss_data[:current_vehicles]
-        service_vehicle_assignments = ss_data[:service_vehicle_assignments] # no empties_or_fills in here
+        return false if ss_data[:cannot_split_further]
 
-        current_vehicles.size > 1 &&
-          (ss_data[:current_vehicle_limit].nil? || ss_data[:current_vehicle_limit] > 1) &&
-          current_vehicles.sum{ |v| service_vehicle_assignments[v.id].size } > vrp.preprocessing_max_split_size
+        # filter the vehicles that are forced to be on the same side
+        current_vehicle_ids = ss_data[:current_vehicles].map(&:id)
+        ss_data[:representative_vrp].relations.each{ |rel|
+          raise 'There should be only :same_route relations inside the representative_vrp.' if rel.type != :same_route
+
+          current_vehicle_ids << rel.linked_ids.join if current_vehicle_ids.reject!{ |id| rel.linked_ids.include?(id) }
+        }
+
+        current_vehicle_ids.size > 1 && (ss_data[:current_vehicle_limit] || Helper.fixnum_max) > 1 &&
+          ss_data[:current_vehicles].sum{ |vehicle|
+            ss_data[:service_vehicle_assignments][vehicle.id].size
+          } > vrp.preprocessing_max_split_size
       end
     end
 
-    # TODO: 0- see below, there are multiple
-    # TODO: 1- implement a notion of "same_vehicle" relation inside balanced_vrp_clustering gem
+    # TODO: Following are ways to improve split_solve performance (2, 3, 4 cpu time) (5 solution quality)
     # TODO: 2- decrease iteration-complexity by "relaxing" the convergence limits (movement, balance-violation) of k-means for max_split?
     # TODO: 3- decrease point-complexity by "grouping" by lat/lon more aggressively for this split
     # TODO: 4- decrease vehicle-complexity by improving balanced_vrp_clustering
@@ -160,9 +183,6 @@ module Interpreters
       # Initialize by first split_by_vehicle and keep the assignment info (don't generate the sub-VRPs yet)
       empties_or_fills = vrp.services.select{ |s| s.quantities.any?(&:fill) || s.quantities.any?(&:empty) }
       vrp.services -= empties_or_fills
-      # TODO: 0- relations needs to be taken into account inside clustering during this split
-      #          - the services need to be in the same route (and by extension, in the same sub-vrp) for the following relations:
-      #            => same_route, order, sequence, shipment
       split_by_vehicle = split_balanced_kmeans(service_vrp, vrp.vehicles.size, cut_symbol: :duration, restarts: 2, build_sub_vrps: false)
 
       # ss_data
@@ -209,10 +229,18 @@ module Interpreters
         sides = split_balanced_kmeans(
           { vrp: create_representative_sub_vrp(ss_data) }, 2,
           cut_symbol: :duration, restarts: 3, build_sub_vrps: false, basic_split: true, group_points: false
-        ).sort_by!{ |side| [side.size, side.sum(&:visits_number)] }.reverse!
-        sides.collect!{ |side| enum_current_vehicles.select{ |v| side.any?{ |s| s.id == v.id } } }
+        ).sort_by!{ |side|
+          [side.size, side.sum(&:visits_number)] # [number_of_vehicles, number_of_visits]
+        }.reverse!.collect!{ |side|
+          enum_current_vehicles.select{ |v| side.any?{ |s| s.id == v.id } }
+        }
 
-        log 'There should be exactly two clusters in split_solve_core!', level: :warn unless sides.size == 2 && sides.none?(&:empty?)
+        unless sides.size == 2 && sides.none?(&:empty?)
+          # this might happen under certain cases (skills etc can force all points to be on one side)
+          # and not necessarily a problem but it should happen very rarely (in real instances)
+          log 'There should be exactly two clusters in split_solve_core!', level: :warn
+          ss_data[:cannot_split_further] = true
+        end
 
         split_service_counts = sides.collect{ |current_vehicles|
           current_vehicles.sum{ |v| ss_data[:service_vehicle_assignments][v.id].size }
@@ -225,13 +253,13 @@ module Interpreters
 
           ss_data[:current_vehicles] = side
 
-          vehicle_limit_ratio = current_vehicle_limit.to_f * side.size / enum_current_vehicles.size
+          v_limit = current_vehicle_limit.to_f * side.size / enum_current_vehicles.size
           # Warning: round does not work if there is an even "half" split
-          ss_data[:current_vehicle_limit] = current_vehicle_limit &&
-            (index.zero? ? vehicle_limit_ratio.ceil : vehicle_limit_ratio.floor)
+          ss_data[:current_vehicle_limit] = current_vehicle_limit && (index.zero? ? v_limit.ceil : v_limit.floor)
 
           split_solve_core(service_vrp, job = nil, &block)
         }
+        ss_data[:cannot_split_further] = false
         log "<-- split_solve_core level: #{split_level}"
 
         Helper.merge_results(results)
@@ -270,14 +298,8 @@ module Interpreters
       sub_vrp.services = ss_data[:current_vehicles].flat_map{ |v| ss_data[:service_vehicle_assignments][v.id] }
       sub_vrp.services.concat ss_data[:transferred_empties_or_fills]
 
-      # fake-shipments are still inside service_vehicle_assignments as services should convert them back
-      # TODO: 0 - With P&D-clustering implementation - Stop treating the "fake-shipments" specially when we support
-      #           shipments in a generic way. The mix of "fake" and "real" shipments will be hard to keep track
-      sub_vrp.shipments = o_vrp.shipments.select{ |ship| sub_vrp.services.reject!{ |ser| ser.id == ship.id } }
-
       # only necessary points -- because compute_matrix doesn't check the difference
       sub_vrp.points = sub_vrp.services.map{ |s| s.activity.point } |
-                       sub_vrp.shipments.flat_map{ |s| [s.pickup.point, s.delivery.point] } |
                        sub_vrp.vehicles.flat_map{ |v| [v.start_point, v.end_point].compact }
 
       # only necessary relations
@@ -298,7 +320,7 @@ module Interpreters
       sub_vrp.configuration = Oj.load(Oj.dump(o_vrp.config)) # time and other limits are correct below
       # split the limits
       sub_vrp.resolution_vehicle_limit = ss_data[:current_vehicle_limit] + ss_data[:transferred_vehicle_limit] if ss_data[:current_vehicle_limit]
-      ratio = (sub_vrp.services.size + 2 * sub_vrp.shipments.size).to_f / (o_vrp.services.size + 2 * o_vrp.shipments.size)
+      ratio = sub_vrp.services.size.to_f / o_vrp.services.size
       sub_vrp.resolution_duration = (o_vrp.resolution_duration * ratio + ss_data[:transferred_time_limit]).ceil if o_vrp.resolution_duration
       sub_vrp.resolution_minimum_duration = (o_vrp.resolution_minimum_duration || o_vrp.resolution_initial_time_out)&.*(ratio)&.ceil
       sub_vrp.resolution_iterations_without_improvment = o_vrp.resolution_iterations_without_improvment&.*(ratio)&.ceil
@@ -309,24 +331,16 @@ module Interpreters
     end
 
     def self.create_representative_vrp(split_solve_data)
-      # This VRP represent the original VRP only `m` number of points by reducing the services belonging to the
+      # This VRP represent the original VRP with only `m` number of points by reducing the services belonging to the
       # same vehicle-zone to a single point (with average lat/lon and total duration/visits). Where `m` is the
-      # number of non-empty vehicle-zones coming from the very first split_by_vehicle.
+      # number of non-empty vehicle-zones generated by the very first split_by_vehicle.
+
       points = []
       services = []
-      # TODO: 0- relations needs to be taken into account inside clustering during this split
-      #          - the vehicles need to be in the same sub-vrp for the following relations:
-      #            => vehicle_trips
-      #          - the services need to be in the same sub-vrp for the following relations:
-      #            => meetup, minimum_duration_lapse, maximum_duration_lapse, minimum_day_lapse, maximum_day_lapse
-      #            (we need to go thorugh the original relations and "connect" the "vehicle_id"s below with "same_route")
-      relations = []
-
       split_solve_data[:service_vehicle_assignments].each{ |vehicle_id, vehicle_services|
-        # TODO: After relations are taken into account inside clustering, we don't have to
-        #       decrease the number of points to 1. We can represent each group with multiple
+        # TODO: We don't have to represent each group with only 1 point. We can represent each group with multiple
         #       points, carefully selected to represent the mean, median and extremes of the group
-        #       and "relate" these points so that they will stay on the same "side" in the 2-split
+        #       and "relate" these points with :same_route so that they will stay on the same "side" in the 2-split
         average_lat = vehicle_services.sum{ |s| s.activity.point.location.lat } / vehicle_services.size.to_f
         average_lon = vehicle_services.sum{ |s| s.activity.point.location.lon } / vehicle_services.size.to_f
         points << { id: "p#{vehicle_id}", location: { lat: average_lat, lon: average_lon }}
@@ -338,6 +352,26 @@ module Interpreters
             duration: vehicle_services.sum{ |s| s.activity.duration.to_f } / Math.sqrt(vehicle_services.size)
           }
         }
+      }
+
+      relations = []
+      # go through the original relations and force the services and vehicles to stay in the same sub-vrp if necessary
+      split_solve_data[:original_vrp].relations.select{ |r| FORCING_RELATIONS.include?(r.type) }.each{ |relation|
+        if relation.linked_vehicle_ids.any? && relation.linked_services.none?
+          relations << { type: :same_route, linked_ids: relation.linked_vehicle_ids }
+        elsif relation.linked_vehicle_ids.none? && relation.linked_services.any?
+          linked_ids = []
+          relation.linked_services.each{ |linked_service|
+            split_solve_data[:service_vehicle_assignments].any?{ |v_id, v_services|
+              linked_ids << v_id if v_services.include?(linked_service)
+            }
+          }
+          linked_ids.uniq!
+          relations << { type: :same_route, linked_ids: linked_ids } if linked_ids.size > 1
+        else
+          # This shouldn't be possible
+          raise 'Unknown relation case in create_representative_vrp. If there is a new relation, update this function'
+        end
       }
 
       # TODO: The following two "fake" vehicles can have carefully selected start and end points!
@@ -378,16 +412,10 @@ module Interpreters
 
         (
           relation.linked_vehicle_ids.empty? ||
-            relation.linked_vehicle_ids.any?{ |linked_v_id|
-              vrp.vehicles.any?{ |v| v.id == linked_v_id }
-            }
+            relation.linked_vehicle_ids.any?{ |id| vrp.vehicles.any?{ |v| v.id == id } }
         ) && (
           relation.linked_ids.empty? ||
-            relation.linked_ids.any?{ |linked_s_id|
-              vrp.services.any?{ |s| linked_s_id = s.id } ||
-              vrp.shipments.any? { |s| linked_s_id == "#{s.id}delivery" } ||
-              vrp.shipments.any? { |s| linked_s_id == "#{s.id}pickup" }
-            }
+            relation.linked_ids.any?{ |id| vrp.services.any?{ |s| s.id == id } }
         )
       }
     end
@@ -529,7 +557,7 @@ module Interpreters
     end
 
     # TODO: private method, reduce params
-    def self.kmeans_process(nb_clusters, data_items, unit_symbols, limits, options = {}, &block)
+    def self.kmeans_process(nb_clusters, data_items, related_item_indices, limits, options = {}, &block)
       biggest_cluster_size = 0
       clusters = []
       restart = 0
@@ -552,7 +580,11 @@ module Interpreters
         # TODO: move the creation of data_set to the gem side GEM should create it if necessary
         options[:seed] = rand(1234567890) # gem does not initialise the seed randomly
         log "BalancedVRPClustering is launched with seed #{options[:seed]}"
-        c.build(DataSet.new(data_items: c.centroid_indices.empty? ? data_items : data_items.dup), options[:cut_symbol], ratio, options)
+        c.build(DataSet.new(data_items: Marshal.load(Marshal.dump(data_items))),
+                options[:cut_symbol],
+                Oj.load(Oj.dump(related_item_indices)),
+                ratio,
+                options)
 
         c.clusters.delete([])
         values = c.clusters.collect{ |cluster| cluster.data_items.collect{ |i| i[3][options[:cut_symbol]] }.sum.to_i }
@@ -615,7 +647,6 @@ module Interpreters
       if vrp.shipments.all?{ |shipment| shipment&.pickup&.point&.location && shipment&.delivery&.point&.location } &&
          vrp.services.all?{ |service| service&.activity&.point&.location } && nb_clusters > 1
         cumulated_metrics = Hash.new(0)
-        unit_symbols = vrp.units.collect{ |unit| unit.id.to_sym } << :duration << :visits
 
         if options[:entity] == :work_day || !vrp.matrices.empty?
           vrp.compute_matrix if vrp.matrices.empty?
@@ -623,7 +654,7 @@ module Interpreters
           options[:distance_matrix] = vrp.matrices[0][:time]
         end
 
-        data_items, cumulated_metrics, linked_objects = collect_data_items_metrics(vrp, cumulated_metrics, options)
+        data_items, cumulated_metrics, grouped_objects, related_item_indices = collect_data_items_metrics(vrp, cumulated_metrics, options)
 
         limits = { metric_limit: centroid_limits(vrp, nb_clusters, data_items, cumulated_metrics, options[:cut_symbol], options[:entity]) } # TODO : remove because this is computed in gem. But it is also needed to compute score here. remove cumulated_metrics at the same time
 
@@ -635,13 +666,13 @@ module Interpreters
 
         options[:clusters_infos] = collect_cluster_data(vrp, nb_clusters)
 
-        clusters = kmeans_process(nb_clusters, data_items, unit_symbols, limits, options, &block)
+        clusters = kmeans_process(nb_clusters, data_items, related_item_indices, limits, options, &block)
 
         toc = Time.now
 
         result_items = clusters.collect{ |cluster|
           cluster.data_items.flat_map{ |i|
-            linked_objects[i[2]]
+            grouped_objects[i[2]]
           }
         }
         log "Balanced K-Means (#{toc - tic}sec): split #{result_items.sum(&:size)} activities into #{result_items.map(&:size).join(' & ')}"
@@ -676,7 +707,7 @@ module Interpreters
         max_cut_metrics = Hash.new(0)
         cumulated_metrics = Hash.new(0)
 
-        data_items, cumulated_metrics, linked_objects = collect_data_items_metrics(vrp, cumulated_metrics)
+        data_items, cumulated_metrics, grouped_objects = collect_data_items_metrics(vrp, cumulated_metrics)
 
         c = AverageTreeLinkage.new
         start_timer = Time.now
@@ -721,7 +752,7 @@ module Interpreters
 
         clusters.delete([])
         result_items = clusters.delete_if{ |cluster| cluster.data_items.empty? }.collect{ |i|
-          linked_objects[i[2]]
+          grouped_objects[i[2]]
         }.flatten
 
         log "Hierarchical Tree (#{end_timer - start_timer}sec): split #{data_items.size} into #{clusters.collect{ |cluster| cluster.data_items.size }.join(' & ')}"
@@ -856,33 +887,11 @@ module Interpreters
         true # if not, they are compatible
       end
 
-      def build_service_like(shipment, activity)
-        Models::Service.new(id: shipment.id,
-                            priority: shipment.priority,
-                            exclusion_cost: shipment.exclusion_cost,
-                            skills: shipment.skills,
-                            activity: activity,
-                            sticky_vehicles: shipment.sticky_vehicles,
-                            quantities: shipment.quantities)
-      end
-
-      def build_services_from_shipments(depot_ids, shipments)
-        shipments.map{ |shipment|
-          if depot_ids.include?(shipment.pickup.point.id)
-            build_service_like(shipment, shipment.delivery)
-          elsif depot_ids.include?(shipment.delivery.point.id)
-            build_service_like(shipment, shipment.pickup)
-          end
-        }.compact
-      end
-
       def collect_data_items_metrics(vrp, cumulated_metrics, options)
-        infeasible_data_items = false
         data_items = []
-        linked_objects = {}
+        grouped_objects = {}
 
         vehicle_units = vrp.vehicles.collect{ |v| v.capacities.to_a.collect{ |capacity| capacity.unit.id } }.flatten.uniq
-        depot_ids = vrp.vehicles.collect{ |vehicle| [vehicle.start_point_id, vehicle.end_point_id] }.flatten.compact.uniq
 
         decimal = if !vrp.matrices.empty? && !vrp.matrices[0][:distance]&.empty? # If there is a matrix, zip_dataitems will be called so no need to group by lat/lon aggresively
                     {
@@ -896,26 +905,26 @@ module Interpreters
                     }
                   end
 
-        # TODO: 0- this part needs to be able to handle real shipments
-        custom_shipments = build_services_from_shipments(depot_ids, vrp.shipments)
+        # TODO: this raise can be deleted once the Models::Shipment is replaced with Relations
+        raise UnsupportedProblemError.new('Clustering supports `Shipments` only as `Relations`') if vrp.shipments.any?
 
-        (vrp.services + custom_shipments).group_by{ |s|
+        vrp.services.group_by{ |s|
           location =
             if s.activity
               s.activity.point.location
             elsif s.activities.size.positive?
-              raise UnsupportedProblemError, 'Clustering is not supported yet if one service has serveral activities.'
+              raise UnsupportedProblemError, 'Clustering does not support services with multiple activities.'
             end
 
-          # TODO: 0- this part needs to be able to handle shipment as a relation
+          can_be_grouped = options[:group_points] && s.relations.none?{ |r| LINKING_RELATIONS.include?(r.type) }
           {
             lat: location.lat.round_with_steps(decimal[:digits], decimal[:steps]),
             lon: location.lon.round_with_steps(decimal[:digits], decimal[:steps]),
             v_id: s[:sticky_vehicle_ids].to_a |
-              [vrp.routes.find{ |r| r.mission_ids.include? s.id }&.vehicle_id].compact,
+              [vrp.routes.find{ |r| r.mission_ids.include? s.id }&.vehicle_id].compact, # split respects initial routes
             skills: s.skills.to_a.dup,
             day_skills: compute_day_skills(s.activity&.timewindows),
-            id: options[:group_points] ? nil : s.id
+            do_not_group: can_be_grouped ? nil : s.id, # use the ID to prevent grouping
           }
         }.each_with_index{ |(characteristics, sub_set), sub_set_index|
           unit_quantities = Hash.new(0)
@@ -938,22 +947,35 @@ module Interpreters
 
           point = sub_set[0].activity.point
           characteristics[:matrix_index] = point[:matrix_index] if !vrp.matrices.empty?
-          linked_objects["#{point.id}_#{sub_set_index}"] = sub_set
+          grouped_objects["#{point.id}_#{sub_set_index}"] = sub_set
           # TODO : group sticky and skills (in expected characteristics too)
           characteristics[:duration_from_and_to_depot] = [0, 0] if options[:basic_split]
           data_items << [point.location.lat, point.location.lon, "#{point.id}_#{sub_set_index}", unit_quantities, characteristics, nil]
         }
 
-        log 'There are services in clustering which cannot be served by any vehicles.', level: :warn if infeasible_data_items
-
-        zip_dataitems(vrp, data_items, linked_objects) if options[:group_points] && vrp.matrices.any? && vrp.matrices[0][:distance]&.any?
+        zip_dataitems(vrp, data_items, grouped_objects) if options[:group_points] && vrp.matrices.any? && vrp.matrices[0][:distance]&.any?
 
         add_duration_from_and_to_depot(vrp, data_items) if !options[:basic_split]
 
-        [data_items, cumulated_metrics, linked_objects]
+        [data_items, cumulated_metrics, grouped_objects, collect_related_item_indices(data_items, grouped_objects)]
       end
 
-      def zip_dataitems(vrp, items, linked_objects)
+      def collect_related_item_indices(data_items, grouped_objects)
+        related_item_indices = Hash.new { |h, k| h[k] = [] }
+
+        data_items.each_with_index{ |data_item, index|
+          sub_set = grouped_objects[data_item[2]]
+          next unless sub_set.size == 1 # linking_relations are not grouped with others
+
+          sub_set.first.relations.select{ |r| LINKING_RELATIONS.include?(r.type) }.each{ |relation|
+            related_item_indices[relation] << index
+          }
+        }
+
+        related_item_indices.group_by{ |k, _v| k.type }.transform_values!{ |v| v.collect!{ |i| i[1] } }
+      end
+
+      def zip_dataitems(vrp, items, grouped_objects)
         vehicle_characteristics = generate_expected_characteristics(vrp.vehicles)
 
         compatible_vehicles = Hash.new{}
@@ -967,7 +989,10 @@ module Interpreters
 
         c.distance_function = lambda do |data_item_a, data_item_b|
           # If there is no vehicle that can serve both points at the same time, make sure they are not merged
-          return max_distance + 1 if (compatible_vehicles[data_item_a[2]] & compatible_vehicles[data_item_b[2]]).empty?
+          if data_item_a[4][:do_not_group] || data_item_b[4][:do_not_group] ||
+             (compatible_vehicles[data_item_a[2]] & compatible_vehicles[data_item_b[2]]).empty?
+            return max_distance + 1
+          end
 
           [
             vrp.matrices[0][:distance][data_item_a[4][:matrix_index]][data_item_b[4][:matrix_index]],
@@ -978,15 +1003,15 @@ module Interpreters
         # Cluster points closer than max_distance to eachother
         clusterer = c.build(DataSet.new(data_items: items), max_distance)
 
-        # Correct items and linked_objects
+        # Correct items and grouped_objects
         clusterer.clusters.each{ |cluster|
           next unless cluster.data_items.size > 1 # Didn't merge any items
 
           item0 = items[items.index(cluster.data_items[0])]
 
           cluster.data_items[1..-1].each{ |d_i|
-            # Merge linked_objects
-            linked_objects[item0[2]].concat(linked_objects.delete(d_i[2]))
+            # Merge grouped_objects
+            grouped_objects[item0[2]].concat(grouped_objects.delete(d_i[2]))
 
             # Merge items
             index_i = items.index(d_i)